refactor: separate PR dashboard rendering from GitHub API utilities
