+++ .github/scripts/bot-on-pr-merged.js
@@ -16,7 +16,7 @@ const {
   postComment 
 } = require('./helpers');
 const { checkMergeConflict } = require('./helpers/checks');
-const { MARKER, buildMergeConflictNotificationComment } = require('./helpers/comments');
+const { PR_HELPER_MARKER, buildMergeConflictNotificationComment } = require('./helpers/comments');
 
 const logger = createLogger('on-pr-merged');
 
@@ -44,7 +44,7 @@ async function checkSiblingConflictsOnMerge(globalBotContext) {
     const actuallyHasConflict = !mergeResult.passed;
 
     // Determine current dashboard status shown to the user
-    const existingComment = await getBotComment(prContext, MARKER);
+    const existingComment = await getBotComment(prContext, PR_HELPER_MARKER);
     const currentlyShowsConflict = existingComment
       ? existingComment.body.includes(':x: **Merge Conflicts**')
       : false;
