+++ .github/scripts/helpers/markers.js
@@ -1,0 +1,12 @@
+// SPDX-License-Identifier: Apache-2.0
+//
+// helpers/markers.js
+//
+// Shared HTML comment markers used for bot comment
+// discovery, deduplication, and updates.
+
+const PR_HELPER_MARKER = '<!-- pr-helper-dashboard -->';
+
+module.exports = {
+  PR_HELPER_MARKER,
+};
